Add functiondiscovery to default reconcilers

+++ porch/scripts/create-deployment-blueprint.sh
@@ -200,9 +200,11 @@ function main() {
 
   IFS=',' read -ra RECONCILERS <<< "$ENABLED_RECONCILERS"
   for i in "${RECONCILERS[@]}"; do
-    # Copy over the CRD
-    cp "${PORCH_DIR}/controllers/config/crd/bases/config.porch.kpt.dev_${i}.yaml" \
-    "${DESTINATION}/0-${i}.yaml"
+    if [[ -f "${PORCH_DIR}/controllers/config/crd/bases/config.porch.kpt.dev_${i}.yaml" ]]; then
+      # Copy over the CRD (if it exists)
+      cp "${PORCH_DIR}/controllers/config/crd/bases/config.porch.kpt.dev_${i}.yaml" \
+         "${DESTINATION}/0-${i}.yaml"
+    fi
     # Update the porch-controllers Deployment env variables to enable the reconciler.
     customize-container-env \
       "ENABLE_${i^^}" \
